+++ src/main/java/it/its/testEmployeesDB/repository/CitiesRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import it.its.testEmployeesDB.dao.CitiesDao;
 
-public interface CitiesRepository extends JpaRepository<CitiesDao, Long> {
+public interface CitiesRepository extends JpaRepository<CitiesDao, Integer> {
 
 }
