+++ src/main/java/it/its/testEmployeesDB/repository/EmployeesRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import it.its.testEmployeesDB.dao.EmployeesDao;
 
-public interface EmployeesRepository extends JpaRepository<EmployeesDao, Long>{
+public interface EmployeesRepository extends JpaRepository<EmployeesDao, Integer>{
 
 }
