+++ src/main/java/it/its/testEmployeesDB/repository/RegionsRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import it.its.testEmployeesDB.dao.RegionsDao;
 
-public interface RegionsRepository extends JpaRepository<RegionsDao, Long> {
+public interface RegionsRepository extends JpaRepository<RegionsDao, Integer> {
 
 }
